Completed Food input functionality and fixed file and folder compositions
