+++ src/components/customer-page/listView.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import '../assets/css/listView.css'
-import greenTimer from '../assets/images/greenTime.png';
-import redTimer from '../assets/images/redTime.png';
-import yellowTimer from '../assets/images/yellowTime.png';
-import addButton from '../assets/images/plus.png';
+import '../../assets/css/listView.css'
+import greenTimer from '../../assets/images/greenTime.png';
+import redTimer from '../../assets/images/redTime.png';
+import yellowTimer from '../../assets/images/yellowTime.png';
+import addButton from '../../assets/images/plus.png';
 
 class ListView extends Component {
     // constructor(props) {
